+++ lib/slack/web/faraday/connection.rb
@@ -13,6 +13,11 @@ module Slack
           options[:proxy] = proxy if proxy
           options[:ssl] = { ca_path: ca_path, ca_file: ca_file }
 
+          request_options = {}
+          request_options[:timeout] = timeout if timeout
+          request_options[:open_timeout] = open_timeout if open_timeout
+          options[:request] = request_options if request_options.any?
+
           ::Faraday::Connection.new(endpoint, options) do |connection|
             connection.use ::Faraday::Request::Multipart
             connection.use ::Faraday::Request::UrlEncoded